Refactor importaciones para utilizar esquemas de URLs en lugar de modelos, y agrega archivo __init__.py en el directorio de esquemas.
