+++ app/utils/external_apis/threatfox.py
@@ -8,7 +8,7 @@ Utiliza la configuración de la aplicación para obtener la clave API y la URL b
 from app.config import settings
 import httpx
 
-from app.models.urls import URLResponse
+from app.schemas.urls_check import URLResponse
 from app.utils.external_api_models.threatfox_model import ThreatFoxResponse
 
 THREATFOX_API_KEY = settings.THREATFOX_API_KEY
